split up StorefrontApiClient
