+++ samples/MobileBuyIntegration/app/src/main/java/com/shopify/checkout_sdk_mobile_buy_integration_sample/di/AppModule.kt
@@ -30,7 +30,8 @@ import com.shopify.buy3.GraphClient
 import com.shopify.buy3.Storefront
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.BuildConfig
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.cart.CartViewModel
-import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.client.StorefrontClient
+import com.shopify.checkout_sdk_mobile_buy_integration_sample.cart.data.source.network.CartStorefrontApiClient
+import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.client.StorefrontApiRequestExecutor
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.logs.LogDatabase
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.logs.Logger
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.logs.MIGRATION_1_2
@@ -38,7 +39,9 @@ import com.shopify.checkout_sdk_mobile_buy_integration_sample.home.HomeViewModel
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.logs.LogsViewModel
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.ProductsViewModel
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.collection.CollectionViewModel
+import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.collection.data.source.network.ProductCollectionsStorefrontApiClient
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.product.ProductViewModel
+import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.product.data.source.network.ProductsStorefrontApiClient
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.settings.PreferencesManager
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.settings.SettingsViewModel
 import kotlinx.coroutines.CoroutineScope
@@ -61,12 +64,20 @@ fun setupDI(application: Application) {
 val appModules = module {
     // App-wide components
     singleOf(::PreferencesManager)
-    singleOf(::StorefrontClient)
+
+    singleOf(::CartStorefrontApiClient)
+    singleOf(::ProductsStorefrontApiClient)
+    singleOf(::ProductCollectionsStorefrontApiClient)
     single {
-        GraphClient.build(
-            context = get(),
-            accessToken = BuildConfig.storefrontAccessToken,
-            shopDomain = BuildConfig.storefrontDomain
+        val maxCacheEntries = 100
+
+        StorefrontApiRequestExecutor(
+            lruCache = LruCache<String, GraphCallResult.Success<Storefront.QueryRoot>>(maxCacheEntries),
+            client = GraphClient.build(
+                context = get(),
+                accessToken = BuildConfig.storefrontAccessToken,
+                shopDomain = BuildConfig.storefrontDomain
+            )
         )
     }
