+++ samples/MobileBuyIntegration/app/src/main/java/com/shopify/checkout_sdk_mobile_buy_integration_sample/products/ProductsViewModel.kt
@@ -25,13 +25,13 @@ package com.shopify.checkout_sdk_mobile_buy_integration_sample.products
 import androidx.lifecycle.ViewModel
 import androidx.navigation.NavController
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.client.ProductPagingSource
-import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.client.StorefrontClient
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.navigation.Screen
+import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.product.data.source.network.ProductsStorefrontApiClient
 import com.shopify.graphql.support.ID
 import timber.log.Timber
 
 class ProductsViewModel(
-    client: StorefrontClient
+    client: ProductsStorefrontApiClient,
 ) : ViewModel() {
 
     val pagingSource = ProductPagingSource(client)
